Add rev.0 hardware schematics
Errata: The microcontroller board needs to be placed
'upside-down', i.e. not as implied by the silkscreen
on the bottom layer of the board.
If in doubt, get in touch and we can verify the placement
together.
The next revision will address this issue. Comitting this
for historical reasons as it was the one actually manufactured
by me.